Revert "Landing page ryan"
